Move templates under templates subfolder
